+++ Modules/Intent.Modules.CosmosDB/release-notes.md
@@ -1,6 +1,7 @@
 ### Version 1.2.4
 
 - Improvement: Added setting to allow enum values to be serialized to a Document as string.
+- Fixed: Value Objects can also now be set in Attribute return types.
 
 ### Version 1.2.3